Extracted unit tests of importer.import_file to a different test suite file to shorten test names
